fix wrong ai impl

+++ gdcef/src/input.rs
@@ -29,7 +29,7 @@ macro_rules! keyboard_modifiers {
 }
 
 /// Extracts mouse button modifier flags from a button mask
-fn mouse_button_modifiers(button_mask: MouseButtonMask) -> i32 {
+fn mouse_button_modifiers(button_mask: MouseButtonMask) -> u32 {
     let mut modifiers = cef_event_flags_t::EVENTFLAG_NONE.0;
 
     if button_mask.is_set(MouseButtonMask::LEFT) {
@@ -55,7 +55,11 @@ pub fn create_mouse_event(
     let x = (position.x * pixel_scale_factor / device_scale_factor) as i32;
     let y = (position.y * pixel_scale_factor / device_scale_factor) as i32;
 
-    MouseEvent { x, y, modifiers: modifiers as u32 }
+    MouseEvent {
+        x,
+        y,
+        modifiers: modifiers as u32,
+    }
 }
 
 /// Handles mouse button events and sends them to CEF browser host
@@ -65,7 +69,8 @@ pub fn handle_mouse_button(
     pixel_scale_factor: f32,
     device_scale_factor: f32,
 ) {
-    let modifiers = keyboard_modifiers!(event) | mouse_button_modifiers(event.get_button_mask());
+    let modifiers =
+        (keyboard_modifiers!(event) | mouse_button_modifiers(event.get_button_mask())) as i32;
     let position = event.get_position();
     let mouse_event =
         create_mouse_event(position, pixel_scale_factor, device_scale_factor, modifiers);
@@ -116,8 +121,12 @@ pub fn handle_mouse_motion(
 ) {
     let modifiers = keyboard_modifiers!(event) | mouse_button_modifiers(event.get_button_mask());
     let position = event.get_position();
-    let mouse_event =
-        create_mouse_event(position, pixel_scale_factor, device_scale_factor, modifiers);
+    let mouse_event = create_mouse_event(
+        position,
+        pixel_scale_factor,
+        device_scale_factor,
+        modifiers as i32,
+    );
     host.send_mouse_move_event(Some(&mouse_event), false as i32);
 }
 
@@ -130,8 +139,12 @@ pub fn handle_pan_gesture(
 ) {
     let modifiers = keyboard_modifiers!(event);
     let position = event.get_position();
-    let mouse_event =
-        create_mouse_event(position, pixel_scale_factor, device_scale_factor, modifiers);
+    let mouse_event = create_mouse_event(
+        position,
+        pixel_scale_factor,
+        device_scale_factor,
+        modifiers as i32,
+    );
 
     let delta = event.get_delta();
     // Convert pan delta to scroll wheel delta
